requirements.txt
Flask==3.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Werkzeug==3.0.3
python-dotenv==1.0.1
bcrypt==4.2.0
